+++ tdesign-miniprogram/switch/switch.d.ts
@@ -1,6 +1,7 @@
 import { SuperComponent } from '../common/src/index';
 export default class Switch extends SuperComponent {
     externalClasses: string[];
+    behaviors: string[];
     properties: import("./type").TdSwitchProps;
     data: {
         classPrefix: string;