+++ tdesign-miniprogram/upload/upload.js
@@ -4,6 +4,17 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
+var __rest = (this && this.__rest) || function (s, e) {
+    var t = {};
+    for (var p in s) if (Object.prototype.hasOwnProperty.call(s, p) && e.indexOf(p) < 0)
+        t[p] = s[p];
+    if (s != null && typeof Object.getOwnPropertySymbols === "function")
+        for (var i = 0, p = Object.getOwnPropertySymbols(s); i < p.length; i++) {
+            if (e.indexOf(p[i]) < 0 && Object.prototype.propertyIsEnumerable.call(s, p[i]))
+                t[p[i]] = s[p[i]];
+        }
+    return t;
+};
 import { isObject, SuperComponent, wxComponent } from '../common/src/index';
 import props from './props';
 import config from '../common/config';
@@ -23,7 +34,6 @@ let Upload = class Upload extends SuperComponent {
             proofs: [],
             customFiles: [],
             customLimit: 0,
-            mediaType: [],
             // 以下是声明properties
             config: {},
             files: [],
@@ -98,30 +108,24 @@ let Upload = class Upload extends SuperComponent {
         this.handleLimit(this.data.customFiles, this.data.max);
         return Promise.resolve();
     }
-    /** 选择图片 */
-    chooseImg() {
-        const { config, max, sizeLimit } = this.data;
-        wx.chooseMedia(Object.assign(Object.assign({ count: max, mediaType: ['image'] }, config), { success: (res) => {
+    /** 选择媒体素材 */
+    chooseMedia(mediaType) {
+        const { config, sizeLimit, max } = this.data;
+        wx.chooseMedia(Object.assign(Object.assign({ count: max === 0 ? 9 : max, // 在 iOS 里，0 是无效的，会导致抛异常
+            mediaType }, config), { success: (res) => {
                 const files = [];
                 res.tempFiles.forEach((temp) => {
-                    if (sizeLimit && temp.size > sizeLimit) {
-                        wx.showToast({ icon: 'none', title: '图片大小超过限制' });
+                    const { size, fileType, tempFilePath, width, height, duration, thumbTempFilePath } = temp, res = __rest(temp, ["size", "fileType", "tempFilePath", "width", "height", "duration", "thumbTempFilePath"]);
+                    if (sizeLimit && size > sizeLimit) {
+                        wx.showToast({ icon: 'none', title: `${fileType === 'image' ? '图片' : '视频'}大小超过限制` });
                         return;
                     }
-                    const name = this.getRandFileName(temp.tempFilePath);
-                    files.push({
-                        name,
-                        type: 'image',
-                        url: temp.tempFilePath,
-                        width: temp.width,
-                        height: temp.height,
-                        size: temp.size,
-                        progress: 0,
-                    });
+                    const name = this.getRandFileName(tempFilePath);
+                    files.push(Object.assign({ name, type: this.getFileType(mediaType, tempFilePath, fileType), url: tempFilePath, size: size, width: width, height: height, duration: duration, thumb: thumbTempFilePath, progress: 0 }, res));
                 });
                 this._trigger('select-change', {
                     files: [...this.data.customFiles],
-                    currentSelectedFiles: [files]
+                    currentSelectedFiles: [files],
                 });
                 this._trigger('add', { files });
                 this._trigger('success', { files: [...this.data.customFiles, ...files] });
@@ -132,68 +136,41 @@ let Upload = class Upload extends SuperComponent {
                 this.triggerEvent('complete', res);
             } }));
     }
-    /** 选择视频 */
-    chooseVideo() {
-        const { config, sizeLimit } = this.data;
-        wx.chooseMedia(Object.assign(Object.assign({ mediaType: ['video'] }, config), { success: (res) => {
-                const files = [];
-                res.tempFiles.forEach((temp) => {
-                    if (sizeLimit && temp.size > sizeLimit) {
-                        wx.showToast({ icon: 'none', title: '视频大小超过限制' });
-                        return;
-                    }
-                    const name = this.getRandFileName(temp.tempFilePath);
-                    files.push({
-                        name,
-                        type: 'video',
-                        url: temp.tempFilePath,
-                        size: temp.size,
-                        width: temp.width,
-                        height: temp.height,
-                        duration: temp.duration,
-                        thumb: temp.thumbTempFilePath,
-                        progress: 0,
-                    });
-                });
-                this._trigger('select-change', {
-                    files: [...this.data.customFiles],
-                    currentSelectedFiles: [files]
-                });
-                this._trigger('add', { files });
-                this._trigger('success', { files: [...this.data.customFiles, ...files] });
-                this.startUpload(files);
-            }, fail: (err) => {
-                this.triggerEvent('fail', err);
-            }, complete: (res) => {
-                this.triggerEvent('complete', res);
-            } }));
+    /**
+     * 由于小程序暂时在ios上不支持返回上传文件的fileType，这里用文件的后缀来判断
+     * @param mediaType
+     * @param tempFilePath
+     * @returns string
+     * @link https://developers.weixin.qq.com/community/develop/doc/00042820b28ee8fb41fc4d0c254c00
+     */
+    getFileType(mediaType, tempFilePath, fileType) {
+        if (fileType)
+            return fileType; // 如果有返回fileType就直接用
+        if (mediaType.length === 1) {
+            // 在单选媒体类型的时候直接使用单选媒体类型
+            return mediaType[0];
+        }
+        // 否则根据文件后缀进行判读
+        const videoType = ['avi', 'wmv', 'mkv', 'mp4', 'mov', 'rm', '3gp', 'flv', 'mpg', 'rmvb'];
+        const temp = tempFilePath.split('.');
+        const postfix = temp[temp.length - 1];
+        if (videoType.includes(postfix.toLocaleLowerCase())) {
+            return 'video';
+        }
+        return 'image';
     }
     // 选中文件之后，计算一个随机的短文件名
     getRandFileName(filePath) {
         const extIndex = filePath.lastIndexOf('.');
         const extName = extIndex === -1 ? '' : filePath.substr(extIndex);
-        return (parseInt(`${Date.now()}${Math.floor(Math.random() * 900 + 100)}`, 10).toString(36) + extName);
+        return parseInt(`${Date.now()}${Math.floor(Math.random() * 900 + 100)}`, 10).toString(36) + extName;
     }
     closePop() {
         this.setData({ showPop: false });
     }
     onAddTap() {
-        const { mediaType } = this.data;
-        // if (mediaType.length > 1) {
-        //   this.setData({ showPop: true });
-        //   return;
-        // }
-        switch (mediaType[0]) {
-            case 'video':
-                this.chooseVideo();
-                break;
-            case 'image':
-                this.chooseImg();
-                break;
-            default:
-                // 后续可能有file等类型
-                break;
-        }
+        const { mediaType } = this.properties;
+        this.chooseMedia(mediaType);
     }
     onChooseImage() {
         this.chooseImg();